soc: add add_constant/add_config methods
